requirements.txt
PyYAML==6.0.2
pandas==2.2.3
SQLAlchemy==2.0.32
ipython==8.26.0